+++ docker-entrypoint.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 
+echo "foo"
 python manage.py migrate
-python manage.py createsuperuser --noinput
+python manage.py setup
 gunicorn -b 0.0.0.0:8000 config.wsgi